Move to src folder and add in strava routes
